simplify, share, move misplaced tests
